server/requirements.txt
Django==1.8
django-extensions==1.5.2
djangorestframework==3.0.0
MySQL-python==1.2.5
six==1.9.0
